VSCODE-292: Show issue links in Jira webview
Also fixed a bug where opening Jira issue webview with issue key was resulting in 404 not found sometimes on Mac due to a race condition between invalidation from abstract webview and the Jira webview implementation. The solution was to make the webview initializer accept only the issue model and not just issue key.