VSCODE-465: Support ssh aliases for Bitbucket features (with some caveats)
Right now we look for exact string `bitbucket.org` in the clone url source to determine if it is a bitbucket repo. When ssh aliases are configured, the source can be set to anything as long as it points to bitbucket.org in the ssh config.
http://blog.developer.atlassian.com/different-ssh-keys-multiple-bitbucket-accounts/

This change loosens the restriction by looking for a string containing `bitbucket.org` instead of exact match, which will allow ssh aliases to work by including that string somewhere in the `Host` value.