VSCODE-517 Added a link to open the issue in the browser.
This is achieved by linking to the issue's base url and adding
/browse and then the issue key. I don't know if the structure of Jira urls
is assumed to be eternal, and therefore can not tell if this is a good
long-term solution. Also, the "Open In Browser" link is currently on a new line,
but perhaps it'd be better to format it so it's on the same line as "Open Issue View".