VSCODE-559 Modified pipeline message for custom pipelines
The window in the bottom-right corner of the screen that pops up when pipelines occur now displays various messages depending on what kind of pipeline it is. If a pipeline is custom, it will display the "custom:<nameOfPipeline> on branch <branchName> <status>". If a custom pipeline is run but there is no branch available, it will display "custom:<nameOfPipeline> <status>". If it was a non-custom pipeline, it will display the old message "<branchName> <status>".

I extended the Pipeline interface so that it could pass more information (like the selector type and pattern), but even so we may eventually need to pass more information. Perhaps the Pipeline interface should be more-or-less equivalent to the structure of the response from the API?

Many different pipelines are now recognized with appropriate text being displayed. These include:
1. Custom pipelines run on a commit
2. Custom pipelines run on a branch
3. Non-custom but still manually triggered pipelines on a commit
4. Non-custom but still manually triggered pipelines on a branch
5. Non-custom, automatically triggered pipelines on a branch

As far as I am aware, this is every type of pipeline that can occur. The way the Pipeline object is currently structured, I am unable to add additional identifying information to the pipeline message in some cases. While I have extended the Pipeline interface, we may want to expand it further in the future to carry even more information.

The pipeline descriptions now use the pipeline number instead of saying "Manual Pipeline" when the name of the pipeline isn't known. Also there was an issue before where when multiple pipelines came in at once, they would still display the "undefined" message. This was because there is a place in the code that checks how many pipelines are coming in and modifies the message depending on the number. They were still referring to the branch name, so I changed that to fit with the style of the single-pipeline message.

When several pipelines come in at once, the message will now only say the name of the pipeline; this excludes the pipeline's status.