VSCODE-593 Authenticating with arbitrary site
When authenticating with a new install for a user who had previously
authed with multiple sites the first site returned is the one that would
be added regardless of which site had been chosen during the OAuth
dance. The solution I came up with is to add all sites associated with
a set of credentials when using them to authenticate. Doing so it made
sense to move from a one site to one keychain entry to one user to one
keychain entry model.