VSCODE-528 Added refresh button to most webviews
There is now a functional refresh button on the pipeline webview,bitbucket issue webview, and pull request webview. In the process, I also fixed a bug with the pull request webview; the refresh command did not refresh comments on PRs.

I am unable to do the Jira webview because I'm unable to get data to load when pressing on an issue. Also, another person is currently changing that page to my knowledge, so it'd be best to wait for those changes to go through before proceeding.