VSCODE-528 Now every time the pull request webview is refreshed, the entire thing is reloaded
Before, pieces of the webview were reloaded, causing some information to either never get reloaded, or get reloaded in fragments. To my knowledge this was done for performance reasons, but now is no longer necessary (it seems to run fine in my testing). The entire page is now loaded at once.