VSCODE-582: Mentions users in bitbucket pullrequests and issues
This is an alpha version for mentioning users. Mentioning someone on bitbucket cloud adds their uuid in the required format in the comment input, so the user will not see the actual user name in the input which is not ideal. It is better than nothing though and we can improve it later.