VSCODE-596 Delete button hides and shows at appropriate times
The delete button on PR comments now only shows up if the comment belongs to the user that is currently authenticated. Additionally, the delete button will not show if the reply text box is open. Part of this fix involved creating a generateActionsList() function which will make adding an edit button later easier. It also makes sure that when the delete button is hidden, there is no empty space following the reply button.