VSCODE-596 Added "edited [time passed] ago" tag to comments that are edited
In order to decide whether to display or not, the creation time and update time of the comment are compared; if there is a difference of more than one second the tag is displayed (equality is not a valid way to check because even unedited comments have a non-zero time difference). Also made it so that empty edits can not be submitted (since this how the website works).