VSCODE-596 Edits and deletes now working on file diff view
There is now an edit and delete icon on comments in the file diff view; these buttons are functional and behave (mostly) like the website. The primary differences are to do with the fact that we have no tree structure in our diff view; also we don't actually refresh the comment list until the refresh button is pressed. This leads to one small issue that I've noticed: when you delete a comment that is the parent of an undeleted comment, the comment will be hidden until the page is refreshed. I don't think this is really a problem, but if it is it can be fixed later (fixing this would require a significant change in the way comment data is currently passed to the PRCommentController).

Additionally, the way deleted comments were hidden (both in webviews and in the diff view) has been changed to happen right when they're acquired from the API. The logic has been altered too to fix a couple bugs resulting from the way it was happening before.