VSCODE-675: Make bitbucket features work with SSH aliases
This feature/workaround to recognize remotes which make use of SSH aliases was inadvertently removed when we switched to sites concept. This adds the feature back. It only works for cloud repos right now. We could look into making it part of site config if there is demand for it in the future.