VSCODE-669 Added diff preview to pull requests
There is now a panel that displays the diff between your most recent commit on your branch and the destination branch. Unfortunately under the hood this is kind of hacky; it's calling and parsing shell scripts. Using the bitbucket api doesn't seem possible because we can't assume that the code has been pushed when the pull request is created. The git api provided by vscode is not sufficient either because their notion of a 'Change' does not contain lines changed.