VSCODE-669 File changes are now only computed from a common fork point
Instead of getting the most recent version of the destination branch, I am computing the diff for the common fork point between destination and source. This ensures that the user will only see changes relevant to the source branch. Additionally, the styles of the changes now better align with the bitbucket website. Finally, the UI was changed such that when there are no results to display, it will stop the loading spinner and instead state that there is nothing to display.