VSCODE-770 Added a few 'hacks' to make searching for relevant pipelines more convenient
If there are still pages left, the program now continues to look through the pages until it finds enough matching pipelines to fill a page, or until it exceeds some search limit. If it finds a pipeline in this process, but not enough to fill the page, it resets it search limit since it has a decent chance of finding this pipeline again.