VSCODE-770 Responded to PR Comments
If the pipelines tree can't find any pipelines matching the filter, it now specifies that it can't find the pipeline between date x to date y. If all dates are searched it will say it can't find any pipelines matching the filter. The behvaior of the tree when there is no filter is unchanged from before.