VSCODE-781 Turned diff list used in PR preview into a component
The primary reason for this is to make it easier to reuse; in particular, I'd like to have such a diff list on the pull request page. As a component, the diff list takes a list of file diffs, a handler function for opening diff views, and some loading variable to know when it should be loading. The display of this data in the format of a file diff list is then handled by the component.