VSCODE-781 All switch statements in our code now have working breaks
Several places in the code had the break statement inside of an if statement, which means if the condition was false then the break would never get executed. In practice, this probably never happened because to trigger that break's case in the first place, you had to be sending data that would satisfy the if statement. Still, since we use those if statements we allow for the possibility that the if condition isn't satisfied, and if that's the case the breaks should be outside of it.