VSCODE-797 Created Onboarding screen for first-time users as specified by Alastair.
There are a few things which are broken (like the text boxes on the custom sites modal, which hopefully will be fixed after merging in devel (I'm assuming this is part of the ESLint stuff). Also, temporarily (for ease of development) I'm checking if the version is undefined or if the version is 0.0.0, but it should only be checking for undefined. It may also be worth changing the place the 'back to vscode' buttons go in the browser, but I'm not sure how important this is.
It is very important that this stuff is tested on other peoples' machines to make sure it behaves as expected, because if there's some strange bugs it'll be the first thing new users see (and we'll get tons of complaints too).