VSCODE-530 Addressed PR Comments and improved performance in various ways, also fixed comment deletion
Performance was improved by:
- Making calls to get tasks and comments concurrent
- Using local state to make tasks appear different when actions are taken on them, even if those actions have not returns API data yet.
   - Deletes now hide the task instantly
   - Adding a task creates a placeholder for the task until it's created
   - Editing a task edits the text of the task and is then confirmed by the API response

Additionally, comment deletions were fixed. It feels like this issue has been "fixed" many times already, but when working on this issue I noticed that comments were still showing up even after they and all their children were deleted.

There are a couple things that should probably still be done, including fixing a minor (but noticable) bug. The bug pertains to clicking the edit button on a task but getting the wrong text inside of it unless you press the edit button again; this is probably fixable with keys. Also, the main 'feature' which should probably still be added is the ability to add inline tasks. I'm still not sure of the best way to approach this...