VSCODE-866 Added ability to create Bitbucket Tasks from the diff view
Comments in the diff view now contain a '+' symbol that will let the user create a new task associated with that comment. This functionality could be extended to allow replies to arbitrary comments, not just the last one of the thread.

This commit/PR hits many files, but that is because in the process of writing this code, I found it would be nicer if all the ids associated with the comment data (so prIds, commentIds, etc.) should be strings. This increases consistency throughout our code, since sometimes ids were represented as strings, and other times as numbers (and also UUIDs are strings).