VSCODE-876 Fixed ordering of BBServer comments
BBServer comments were disordered at the top-level. When replying to comments, comment order was preserved, but adding a new comment caused that comment to appear at the top. This was fixed by adding a sorting step before returning the comment data. If later it is discovered that comment ordering is not preserved in other instances (e.g. when replying to comments or adding tasks), the sorting step can simply be moved to before the toNestedCommentModel step; I chose not to do this for now because it is far more expensive to sort every comment than it is to sort top-level comments (although to be fair, it is unlikely any PR has enough comments for this to be that significant).