VSCODE-868 Added Task List (with PR-Level tasks for BBCloud)
The PR Details page now contains a list of all the tasks collected in one place (this is in a panel and can be expanded). Moreover, you can now add and view PR-level tasks for BBCloud (BBServer doesn't support this).
Part of achieving this required a slight refactor of task actions. Comment Ids are now optional. Additionally, Comment objects are no longer passed around (just IDs).