VSCODE-868 Restyled tasks and fixed some strange behavior by introducing keys
The checkbox of the task now floats to the left while the rest of the element sits to the right of it. Additionally, some issues with spacing were resolved by forcing the padding of an outer div to be 0px (it was defaulting to 28px for some reason).

Also, keys were added to the task list, so their order is preserved after deletes, meaning actions taken on them should apply to the correct elements always now.