VSCODE-902 Fixed bbserver issue where replies weren't showing up until the page was reloaded (diff view)
When a reply was made on the diff view page, the reply wouldn't show up until the page was reopened. This is because though BBServer stores the new comment as being a reply to some other comment (and requires the parent id when posting), it does not send back the comment with the parent id. Therefore, parent id checks were failing, so instead I am now placing the reply after a comment with a defined 'temporaryReply' property (there should only ever be one).