VSCODE-947 JQL in Jira Explorer now show issue counts
This is achieved by executing JQL queries immediately instead of upon expanding. This already greatly improves usability of the explorer due to increased speed. One potential concern was that this mass querying would slow down performance, but this doesn't appear to be the case.

Currently, there appears to be an issue where when the refresh button is pressed, issues are updated but the counts are not. This is because it appears that nodes are unchanging, and need to be destroyed/rebuilt to show changes. I'll look into possible ways of implementing this...

The user who suggested this change also suggested that when expanding the issue node, greater detail would be shown about the issue breakdown (bugs, tasks, epics, etc). This appears to not be possible; there does not appear to be a way to update the node upon expanding it.