VSCODE-947 Corrected issue counts and addressed PR concerns
It turns out that the _issues memeber variable was storing a tree of issues (e.g. Epics), so counting the issues by taking its length was giving an inaccurate count. The issues are now totalled up by traversing the issue tree and counting up all the results, and that result is stored in the _numIssues variable.

Issue counts have been reformatted to display better in more cases (e.g. only one issue or 100 issues when there is an issue cap).