VSCODE-941 Authing on onboarding screen no longer sends to settings, and can now enable/disable bitbucket/jira from onboarding screen
The onboarding screen will now return you to the onboarding screen when 'back to vscode' is pressed after authing. Better still is that since this page is already open, it appears that the alert window from vscode asking if you'd like to open an external link no longer shows up.

Additionally, you can now enable/disable both bitbucket and jira from the onboarding page. This changes what auth options you have available, and also removes the 'continue' button if both Jira and Bitbucket are removed.

This is really part 1 of a multi-part reworking of the onboarding process. I figured it made sense to first address what was asked for in the associated Jira ticket, and then move forward with additional improvements to our onboarding process.