VSCODE-925 Removed state associated with flattened issue lists
instead of storing the issue count and the flattened issue list for each node, executeQuery() now generates this dynamically. This means there is far less memory usage because the same issues don't need to be stored multiple times.

Additionally, a complicated lambda funciton which showed the issue search quick pick is now in its own function.