VSCODE-985 Reconciled issues with Task state vs task props
TaskComponents use state to apply local changes while waiting for an API response to synchronize the data. This allows for a seemless experience for the user, because it creates an illusion of instantaneous actions. However, once all data transmission was separated (tasks, comments, etc. are no longer sent at the same time as basic PR data), the state of the PR page was being updated many different times. Modified state does not necessarily cause a rerender of child components, but as I painfully learned, it DOES cause componentWillReceiveProps to be called for any children that depend on the state of the page (even if that particular state they depend on wasn't modified). To temporarily get around this, I no longer accept props in the TaskComponent unless the given Task prop is actually different than the one the component already has. This is not a real solution and will have to be dealt with eventually, because in the event that the Task API call fails, it won't reset the state of the Task Component. That means the user will think their changes went through when really they didn't.

Most changes in this PR are formatting changes due to Prettier... unfortunately it's very difficult to notice Prettier has done this until it's too late, otherwise I would've at least put those changes in separate commits.