VSCODE-999 Added PR preloading and caching to BBCloud
When the PR is initially loaded, basic data for PRs is gathered and PR nodes are created (with their data being preloaded via a promise) and cached. Any time the refresh button is clicked, basic data is again gathered for the PRs. Additional data must be gathered for PRs whose updated times differ from their cached updated times, but if the PR has the same updated time as before, we can simply return the cached node. This greatly enhances performance, because by the time a user clicks on a PR node it is almost certainly preloaded, and when refresh is clicked only nodes that must be reloaded are reloaded.

Unfortunately, these efforts were only effective in BBCloud. On BBServer, the updated time does not refer to the last time any PR action was taken (e.g. comment, task, etc.). Instead, it only tracks when the description, title, and a few other major things are updated. As a result, we can not cache BBServer PRs because we do not know if they are up to date. Furthermore, rate limits on custom instances may vary, and we can not assume that hundreds of API calls to BBServer are acceptable, so even PR preloading can't be done.