VSCODE-999 Changed the conditions under which preloading occurs
Before, any time a pull request was a cloud pr, it would be preloaded. This runs into potential issues with rate limits for clients with 100s of PRs open, so this is now capped to 150. Additionally, server PRs were never preloaded, but now are preloaded when there are less than 10 of them, because this is very unlikely to cause problems.

Additionally, there is no longer a cache at the class-level in repositoriesNode. Instead, we take advantage of the fact that nodes are basically already cached inside the children array, so we create a map on the spot inside createChildNodes(). Also, nodes are now created in batches instead of individually (e.g. createChildNodes() vs createChildNode()).