VSCODE-1010 Added repo name to pipeline notifications
This is only a one line change, but it warrents some thought because it makes the notifications potentially very long and cluttered if there is more than one pipeline result. Furthermore, most users probably only have one repo, so I'm not sure how great this is. Don't want to over-engineer this either though because I don't think that many users are too interested in pipeline result notifications...