VSCODE-1029 Stop infinite credential refresh
Updating credentials causes clients to be recreated. Recreating OAuth
clients causes then to refresh their access token. Saving the access
token causes the credentials to update...

But, there's a bug in our comparison logic for whether or not
credentials have actually been updated so sometimes (most of the time)
this doesn't happen.