VSCODE-1013 Focus the Jira explorer window after authing with Jira on the Onboarding screen
This is using the reveal() method from the vscode API. Unfortunately, implementing reveal required a lot of changes to our tree structure. What I did amount to patching up a few things, but really we may need to take a closer look at our code and rewrite our trees.

I originally intended to do the same for the PR explorer when you auth with Bitbucket, but this is turning out to be pretty complicated with the way our trees are set up. For now, I think it's best to keep this separate, because that will be its own can of worms...

Also fixed a long-standing type (treeDataProvder -> treeDataProvider)