VSCODE-988 Built most of the UI, backend logic needs to be hooked up
Built most of the onboarding screen UI. The UI is now split into 3 pages:

1. Product selection
2. Authentication
3. A feature demo page

This is all tied together with a Stepper at the top (it may be worth considering a switch to a vertical stepper).
The feature demo page contains gifs with features being demonstrated, and clicking the buttons will eventually perform an action.
UI components were carefully styled for a clean look, but this comes with some caveats. My primary concerns are that the pages look a bit sparse (though this makes the UI somewhat future proof for additional products), and that the styling is very specific in some places so the code isn't too clean. I'm hoping to clean up the code a bit later and maybe a code review will help with that.

The UI logic works (proper page navigation and such), so mostly what remains is to hook up the buttons to the "back end".