VSCODE-1197 Separated fetching of comments and tasks
This slightly complicates logic in a few places, but ultimately is easier to reason about because each respective call to comments and tasks only grabs comments or tasks. Additionally, this should increase speed since comment calls no longer have to wait for tasks calls to finish. It also means comments can be displayed without tasks as soon as the response is available, and the tasks can be populated later.