[AXON-297] Epics do not show child issues (#564)
* Made some progress on addressing epic issues

Looks like this is bigger issue than just not fetching data. Maybe some problems exists with the Jira APIs themselves and we may be using removed API endpoints.

* Changed the query for fetching epic children of issue and adjusted to depericated field value

Changed the JQL query that fetches epic children and manually updated a depreicated field value for an issue, isEpic, as a bandaid fix without modifying the pi-client itself.

* Got the epic children issues to render

Still need to adjust the React props to fit what is needed for Epic children issues. Also need to discuss if we want done issues to show.

* Update the props for epics

Update props to render certain attributes for epic issues.

* Added Functionality to creating epic children

There was no functionality to create epic children. Now the epic children show and have functionality.

* Added Unit tests for changes made in jiraIssueWebview.ts

Made some AI generated enhancements to the current jiraIssueWebview test file to accomidate for changes to epic issue flows.

* Added support for Server instances

There was no support for epics and their children for cloud instances. The cloud uses parent and child relationship while in DC instance we have the old epic and epic child relationship. Epics should now be supported for Cloud instances and DC instances versions 7.0+ (according to rovo) but I tested on 9.11 in instEnv.

* Deleted of some comments